+++ changelog.md
@@ -1,5 +1,6 @@
 ## Current
+- Added asset 'manual'
+- Added PEGI rating
+## Previous
 - Updated dependencies
 - Fix for nplayers metadata
-## Previous
-- Initial release of the MobyGames AKL plugin